Pretty much rewrite everything.
Really should've committed this as I went along but it was all too
messy and none of it worked until very recently.

This is pretty much a chuck-everything-out-the-window-and-rewrite-it job.
Various bits have been kept around but a lot has been redone from scratch.

 * Swap showdown out for markdown-it (which is nicer to use and
   better-maintained than marked)
 * Use various modules like ejs to get rid of lots of stuff done in-house
 * Kill off pygments and use highlight.js. This changes the languages
   available for highlighting, but makes everything waaaaay faster, easier
   to work with, and doesn't require the messy pygments child process stuff.
   It also changes the available colour schemes, which isn't great, sorry :(
 * Some messy stuff that I still want to clean up for grabbing colour schemes
   straight from the highlight.js styles and feeding them into LESS
 * Oh yeah, switch from sass to less for styles because native node = yay

This has been done in pretty much exactly the wrong style: lots of code,
a lot of which *isn't* self-explanatory, and not many comments. Before I
get this into master I'll make sure it's up-to-scratch with comments.